whisper-docker/requirements.txt
openai-whisper
torch
numpy
ffmpeg-python
tk 